Move auth entities to new namespace
