+++ DLCS.Model/Auth/Entities/AuthService.cs
@@ -1,6 +1,6 @@
 ﻿#nullable disable
 
-namespace DLCS.Model.Auth
+namespace DLCS.Model.Auth.Entities
 {
     public partial class AuthService
     {
