+++ DLCS.Model/Auth/IRoleProviderRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace DLCS.Model.Auth
+{
+    public interface IRoleProviderRepository
+    {
+        
+    }
+}
