+++ DLCS.Repository/DlcsContext.cs
@@ -6,6 +6,7 @@ using DLCS.Model.Assets;
 using DLCS.Model.Assets.CustomHeaders;
 using DLCS.Model.Assets.NamedQueries;
 using DLCS.Model.Auth;
+using DLCS.Model.Auth.Entities;
 using DLCS.Model.Customers;
 using DLCS.Repository.Auth;
 using DLCS.Repository.Entities;
