+++ Test.Helpers/Integration/DlcsDatabaseFixture.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using DLCS.Model.Assets;
 using DLCS.Model.Auth;
+using DLCS.Model.Auth.Entities;
 using DLCS.Model.Customers;
 using DLCS.Repository;
 using DLCS.Repository.Entities;
